Changed the ideal- and -Rating variables in EndRatingsCalculator to int type.

+++ PEC-App/Assets/Scripts/EndRatingsCalculator.cs
@@ -14,34 +14,34 @@ public class EndRatingsCalculator : MonoBehaviour
     /// The ideal wall saturation value. Used to determine final wall saturation rating.
     /// </summary>
     [SerializeField]
-    private float m_idealWallSaturation;
+    private int m_idealWallSaturation;
 
     /// <summary>
     /// The ideal air saturation value. Used to determine final air saturation rating.
     /// </summary>
     [SerializeField]
-    private float m_idealAirSaturation;
+    private int m_idealAirSaturation;
 
     /// <summary>
     /// The ideal money spent value. Used to determine final air saturation rating.
     /// </summary>
     [SerializeField]
-    private float m_idealMondaySpent;
+    private int m_idealMondaySpent;
 
     /// <summary>
     /// The final wall saturation rating value.
     /// </summary>
-    private float m_wallSaturationRating = 0;
+    private int m_wallSaturationRating = 0;
 
     /// <summary>
     /// The final air saturation rating value.
     /// </summary>
-    private float m_airSaturationRating = 0;
+    private int m_airSaturationRating = 0;
 
     /// <summary>
     /// The final money spent rating value.
     /// </summary>
-    private float m_moneySpentRating = 0;
+    private int m_moneySpentRating = 0;
 
     public void CalculateEndRatings()
     {
